+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2017-01-27
+- Made the dependencies (json_sh) and configuration modular
+- Added writing a status file upon last successful execution
+- Made some smaller enhancements to ppaas.sh
+
 ## [1.0.1] - 2016-10-27
 - Added Cloudflare API v4 Availability check
 - Fixed the Pingdom Probes Availability check